WebSockets/tests/submissions/Opera/interfaces/CloseEvent/002.html

<!doctype html>
<title>WebSockets: dispatching a CloseEvent</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
setTestTimeout(2000);
test(function(){
  var e = document.createEvent('CloseEvent');
  e.initCloseEvent('foo', false, false, false);
  var div = document.createElement('div');
  div.addEventListener('foo', function(ev) {
    assert_equals(ev.bubbles, false);
    assert_equals(ev.cancelable, false);
    assert_equals(ev.wasClean, false);
    end();
  }, false);
  div.dispatchEvent(e);
});
</script>
